requirements.txt
aiofile==3.8.8
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
attrs==23.1.0
caio==0.9.13
cssselect==1.2.0
diskcache==5.6.3
frozenlist==1.4.0
humanize==4.9.0
idna==3.6
isodate==0.6.1
jmespath==1.0.1
llama_cpp_python==0.2.20
lxml==4.9.3
multidict==6.0.4
numpy==1.26.2
packaging==23.2
parsel==1.8.1
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
w3lib==2.1.2
yarl==1.9.4
